requirements.txt
# === Core ===
Django>=5.2,<5.3
djangorestframework>=3.15,<3.16
djangorestframework-simplejwt>=5.3,<5.4
django-cors-headers>=4.4,<4.5
django-filter>=24.0,<25.0
python-decouple>=3.8,<4.0

# === Sécurité ===
bleach>=6.1,<7.0

# === Médias & Stockage ===
Pillow>=10.4,<11.0
boto3>=1.35,<2.0
django-storages>=1.14,<2.0

# === Serveur de production ===
gunicorn>=22.0,<23.0
whitenoise>=6.7,<7.0

# === Base de données ===
psycopg2-binary>=2.9,<3.0
